projekty/spinoza/pyproject.toml
[project]
name = "spinoza"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "aiohttp>=3.11.14",
    "coverage>=7.7.0",
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.7",
    "django-extensions>=3.2.3",
    "djangorestframework>=3.15.2",
    "factory-boy>=3.3.3",
    "faker>=37.0.1",
    "httpie>=3.2.4",
    "ipython>=9.0.2",
    "markdown>=3.7",
    "pygments>=2.19.1",
    "requests>=2.32.3",
    "ruff>=0.11.1",
    "unicorn>=2.1.3",
    "uvicorn>=0.34.0",
]
